move generate to post voucher
